feat: add Py module crosscheck for testing micro vs lite
Add a Python module `tflite_micro.test_utils.crosscheck` for
testing the micro interpreter against the lite interpreter. The
initial facility is the `versus_lite` function, which accepts any
model, runs both interpreters with the same random input, and
expects equivalent outputs.

This initial implementation may have some limitations in the
characteristics of models it accepts, the degree to which
quantized ranges are uniformly exercised, etc. Improvement over
time is expected. The name crosscheck is general enough to
include testing micro against other interpreters in the future.

Include unit tests which test the module and serve as a simple
example of its use.