feat(compression): add metadata flatbuffer schema and tests
Add a flatbuffer schema for describing compressed models. Flatbuffers
with this schema are to be used as the value in a .tflite model
flatbuffer metadata field, and contain the extra information necessary
to describe a compressed model.

Include tests to ensure basic functionality and demonstrate integration
with C++, Python, and Bazel.