feat(compression): save metadata header library for use by Make
Save the generated header-only library used to access metadata
flatbuffers to the source tree and commit to git. Modify the C++ unit
test to use this saved header. Add a test to ensure consistency between
this saved header and the flatbuffer schema. Add a script for updating
this saved header when the flatbuffer schema changes. Exempt the saved
header from code style checks.

During a Bazel build, the existing target `:metadata_cc` generates, from
the metadata flatbuffer schema `metadata.fbs`, a header-only library for
accessing the metadata flatbuffer. Code built (only, see below) by Bazel
can depend on this target and thereby use the generated header. Our code
built by Make, however, does not have a means of generating the header
from the schema during the build. For the sake of this code, save a copy
of the generated file here to be statically read during the Make build.

Because manually saving a file generated from the schema runs the risk
that the schema is changed but the saved file is not manually updated,
add a test that will fail if the saved file diverges from the schema.
Add a script that can be run manually to update the saved file.

Since most code is buildable from both Bazel and Make, most code must
use this saved header rather than the generated header. In Bazel, this
requires using the ":metadata_saved" target and `#include
".../metadata_saved.h"` rather than the ":metadata_cc" target and
`#include ".../metadata_generated.h".