build(bazel): introduce tflm_cc_* macros, refactoring away micro_copts
Remove micro_copts() by replacing every cc_* target that used
them with a tflm_cc_* equivalent, and setting those common copts in one
place, inside the tflm_cc_* macro.

This is the first of several commits introducing tflm_cc_* macros in
place of cc_binary, cc_library, and cc_test. Motivated by the upcoming
need to support conditional compilation, the objective is to centralize
build configuration rather than requiring (and remembering that) each
cc_* target in the project add the same common attributes such as
compiler options and select()ed #defines.

Alternatives such as setting global options on the command line or in
.bazelrc, even if simplified with a --config option, fail to preserve
flags and hooks for configuration in the case TFLM is used as an
external repository by an application project. Nor is it easy in that
case for individual targets to override an otherwise global setting.